added arguments for the bash script, fixed some tab-space issues and modified the target of registries untar

+++ dspace_metadata_import.sh
@@ -3,7 +3,7 @@
 
 ########################################################################################
 ## Run this script to import metadata from old dspace into new dspace6 database, and  ##
-## copy over the metadata registry files.											  ##
+## copy over the metadata registry files.                                             ##
 ########################################################################################
 
 
@@ -11,23 +11,59 @@
 ## DSpace6 Upgrading: https://wiki.duraspace.org/display/DSDOC6x/Upgrading+DSpace ##
 
 
-#########################################################################
-#                       									 			#
-#	Need to 															#
-#		1) copy over the dumped pgsql file from old dspace, and			#
-#		2) set up the ownership, and									#
-#		3) replace the user to be dspace or something, and 				#
-#		4) comment out the block of adding rule protect_approval_upd	#
-#    	5) comment out the privileges granted to user libacct			#
-#                       												#
-#########################################################################
+#################################################################################
+#                                                                               #
+#       Need to                                                                 #
+#		1) copy over the dumped pgsql file from old dspace, and         #
+#		2) set up the ownership, and                                    #
+#		3) replace the user to be dspace or something, and              #
+#		4) comment out the block of adding rule protect_approval_upd    #
+#               5) comment out the privileges granted to user libacct           #
+#                                                                               #
+################################################################################
 
-PATH_TO_REGISTRIES_FROM_OLD_DSPACE=/tmp/registries.tar.gz
-PATH_TO_OLD_DSPACE_DATABASE_DATA=/tmp/test3.pgsql
-EXISTING_DSPACE_DATABASE_NAME=dspace
-EXISTING_DSPACE_DATABASE_NAME_CHANGED=dspace_old
-NEW_DSPACE_DATABASE_NAME=dspace2
-DSPACE_DATABASE_OWNER=dspace
+
+EXISTING_DSPACE_DATABASE_NAME_CHANGED=dspace_changed_from_existing_temp
+NEW_DSPACE_DATABASE_NAME=dspace_new_temp
+
+while getopts o:p:q:r:s:t: option
+  do
+    case "${option}"
+    in
+      o) PATH_TO_REGISTRIES_FROM_OLD_DSPACE=${OPTARG};;
+      p) PATH_TO_OLD_DSPACE_DATABASE_DATA=${OPTARG};;
+      q) EXISTING_DSPACE_DATABASE_NAME=${OPTARG};;
+      r) DSPACE_DATABASE_OWNER=${OPTARG};;
+      s) DATABASE_HOSTNAME=$OPTARG;;
+    
+ esac
+done
+
+if [[ -z "${PATH_TO_REGISTRIES_FROM_OLD_DSPACE// }" ]]; then
+	echo "ERROR: the path to the registries of the old dspace instance is not defined!"
+	exit
+fi
+
+if [[ -z "${PATH_TO_OLD_DSPACE_DATABASE_DATA// }" ]]; then
+	echo "ERROR: the path to the database data file of the old dspace instance is not defined!"
+	exit
+fi
+
+if [[ -z "${EXISTING_DSPACE_DATABASE_NAME// }" ]]; then
+	echo "ERROR: the name of the database of the existing dspace instance is not defined!"
+	exit
+fi
+
+if [[ -z "${DSPACE_DATABASE_OWNER// }" ]]; then
+	echo "ERROR: the name of the dspace database owner is not defined!"
+	exit
+fi
+
+if [[ -z "${DATABASE_HOSTNAME// }" ]]; then
+	echo "ERROR: the hostname of the dspace database is not defined!"
+	exit
+fi
+
 
 sudo systemctl stop tomcat
 
@@ -49,7 +85,7 @@ EOF
 
 echo "#####  Import metadata into the new database  #####"
 
-sudo -u postgres psql -U $DSPACE_DATABASE_OWNER -h localhost $NEW_DSPACE_DATABASE_NAME < $PATH_TO_OLD_DSPACE_DATABASE_DATA
+sudo -u postgres psql -U $DSPACE_DATABASE_OWNER -h $DATABASE_HOSTNAME $NEW_DSPACE_DATABASE_NAME < $PATH_TO_OLD_DSPACE_DATABASE_DATA
 
 echo "##### Modify the databases #####"
 
@@ -67,7 +103,7 @@ EOF
 
 echo "#####  Copy over the registry files  #####"
 sudo mv /srv/oulib/dspace/config/registries/ /srv/oulib/dspace/config/registries-old/
-sudo tar -xvf $PATH_TO_REGISTRIES_FROM_OLD_DSPACE -C /tmp
+sudo mkdir /tmp/registries && tar -xvf $PATH_TO_REGISTRIES_FROM_OLD_DSPACE -C /tmp/registries --strip-components=2
 sudo mv /tmp/registries /srv/oulib/dspace/config
 sudo chown -R tomcat:tomcat /srv/oulib/dspace/config/registries/
